pyproject.toml
[tool.poetry]
name = "dfa"
version = "0.1.0"
description = ""
authors = ["subbotinav"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
djoser = "^2.1.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
